pyproject.toml
[project]
name = "cv-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "albumentations>=2.0.8",
    "matplotlib>=3.10.6",
    "scikit-learn>=1.7.2",
    "tensorboard>=2.20.0",
    "torch-cpu>=2.8.0",
    "torchvision-cpu>=0.23.0",
    "tqdm>=4.67.1",
]
